move getPaginationMap to base resource class
