+++ src/main/groovy/edu/oregonstate/mist/osuevents/resources/CampusesResource.groovy
@@ -8,7 +8,6 @@ import edu.oregonstate.mist.osuevents.ResourceObjectBuilder
 import edu.oregonstate.mist.osuevents.core.Campus
 import edu.oregonstate.mist.osuevents.core.PaginatedCampuses
 import edu.oregonstate.mist.osuevents.db.LocalistDAO
-import edu.oregonstate.mist.osuevents.db.PaginationObject
 import groovy.transform.TypeChecked
 
 import javax.annotation.security.PermitAll
@@ -63,45 +62,13 @@ class CampusesResource extends Resource {
                 data: paginatedCampuses.campuses.collect {
                     campusResourceObject(it)
                 },
-                links: getPagiationLinks(paginatedCampuses.paginationObject)
+                links: getPagniationLinkMap(resourcePath, paginatedCampuses.paginationObject.total)
         )
 
         ok(resultObject).build()
     }
 
     ResourceObject campusResourceObject(Campus campus) {
-        resourceObjectBuilder.buildResourceObject(campus.campusID, "campuses", campus)
-    }
-
-    private Map getPagiationLinks(PaginationObject paginationObject) {
-        def links = [:]
-
-        links['self'] = getPaginationUrl(['pageNumber': getPageNumber(), 'pageSize': getPageSize()],
-                resourcePath)
-
-        links['first'] = getPaginationUrl(['pageNumber': 1, 'pageSize': getPageSize()],
-                resourcePath)
-
-        links['last'] = getPaginationUrl(['pageNumber': paginationObject.total,
-                                          'pageSize': getPageSize()],
-                resourcePath)
-
-        if (getPageNumber() <= 1) {
-            links['prev'] = null
-        } else {
-            links['prev'] = getPaginationUrl(['pageNumber': getPageNumber() - 1,
-                                              'pageSize': getPageSize()],
-                    resourcePath)
-        }
-
-        if (getPageNumber() >= paginationObject.total) {
-            links['next'] = null
-        } else {
-            links['next'] = getPaginationUrl(['pageNumber': getPageNumber() + 1,
-                                              'pageSize': getPageSize()],
-                    resourcePath)
-        }
-
-        links
+        resourceObjectBuilder.buildResourceObject(campus.campusID, resource, campus)
     }
 }
